fix: resolve PR #1525 review comments
- [service_provider] Add nil checks for Config, Auth, Log, Queue facades in Register
- [errors] Add AuthFacadeNotSet error
- [setup] Restore PUSHER_APP_KEY, PUSHER_APP_SECRET, PUSHER_APP_ID, PUSHER_HOST env vars
- [contracts] Move config structs to contracts/broadcasting/config.go (separate file from interfaces)
